+++ lib/Models/Terria.ts
@@ -351,6 +351,11 @@ export interface ConfigParameters {
    */
   searchBarConfig?: ModelPropertiesFromTraits<SearchBarTraits>;
   searchProviders: ModelPropertiesFromTraits<SearchProviderTraits>[];
+
+  /**
+   * If true search also in info of catalog layers.
+   */
+  searchInCatalogItemInfo: boolean;
 }
 
 interface StartOptions {
@@ -568,7 +573,8 @@ export default class Terria {
     aboutButtonHrefUrl: "about.html",
     plugins: undefined,
     searchBarConfig: undefined,
-    searchProviders: []
+    searchProviders: [],
+    searchInCatalogItemInfo: false
   };
 
   @observable